# docs/api/api_docs.yaml
openapi: 3.0.0
info:
  title: Sync Gateway API
  version: '3.0'
  license:
    name: Business Source License 1.1 (BSL)
    url: 'https://github.com/couchbase/sync_gateway/blob/master/LICENSE'
servers:
  - url: 'http://localhost:4984'
    description: Public API
  - url: 'http://localhost:4985'
    description: Admin API
  - url: 'http://localhost:4986'
    description: Metric API
paths:
...
  /_replicationStatus/:
    get:
      responses:
        '200':
          description: Successfully retrieved all replication statuses.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Replication-status'
        '400':
          $ref: '#/components/responses/request-problem'
      tags:
        - Admin
      summary: Get all replication statuses
      description: Retrieve all the replication statuses in the Sync Gateway node.
      parameters:
        - $ref: '#/components/parameters/replication-active-only'
        - $ref: '#/components/parameters/replication-local-only'
        - $ref: '#/components/parameters/replication-include-error'
        - $ref: '#/components/parameters/replication-include-config'
    head:
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      tags:
        - Admin
  '/_replicationStatus/{replicationid}':
    parameters:
      - $ref: '#/components/parameters/replicationid'
    get:
      responses:
        '200':
          description: Successfully retrieved replication status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Replication-status'
        '400':
          $ref: '#/components/responses/request-problem'
        '404':
          description: Could not find replication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP-Error'
      tags:
        - Admin
      summary: Get replication status
      description: Retrieve the status of a replication.
      parameters:
        - $ref: '#/components/parameters/replication-active-only'
        - $ref: '#/components/parameters/replication-local-only'
        - $ref: '#/components/parameters/replication-include-error'
        - $ref: '#/components/parameters/replication-include-config'
    put:
      responses:
        '200':
          description: Successfully changed target state of replicator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Replication-status'
        '400':
          $ref: '#/components/responses/request-problem'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP-Error'
        '404':
          $ref: '#/components/responses/Not-found'
      tags:
        - Admin
      description: |-
        Change the replication state to target. 

        This is done through the action query parameter, which has 3 valid values:
        * `start` - starts a stopped replication
        * `stop` - stops an active replication
        * `reset` - resets the replication checkpoint to 0. For bidirectional replication, both push and pull checkpoints are reset to 0. The replication must be stopped to use this.
      summary: Modify a replication status
      parameters:
        - name: action
          schema:
            type: string
            enum:
              - start
              - stop
              - reset
          in: query
          description: The target state to put the replicator into.
          required: true
    head:
      responses:
        '200':
          description: Replication exists
        '400':
          $ref: '#/components/responses/request-problem'
        '404':
          $ref: '#/components/responses/Not-found'
      tags:
        - Admin
      summary: Check if replication exists
      parameters:
        - $ref: '#/components/parameters/replication-active-only'
        - $ref: '#/components/parameters/replication-local-only'
        - $ref: '#/components/parameters/replication-include-error'
        - $ref: '#/components/parameters/replication-include-config'
      description: Check if a replication exists
  /_logging:
    get:
      responses:
        '200':
          description: Returned map of console log keys
          content:
            application/json:
              schema:
                type: object
                description: Map of console log keys
                properties:
                  log_key:
                    type: boolean
                    default: true
                    description: This is the console log keys currently in use (therefore all the values will be `true`).
                required:
                  - log_key
      tags:
        - Admin
      description: |-
        **Deprecated in favour of `GET /_config`**
        This will return a map of the log keys being used for the console. This will include the runtime log keys.
      summary: Get the console log keys
      deprecated: true
    put:
      responses:
        '200':
          description: Log keys successfully replace
        '400':
          $ref: '#/components/responses/request-problem'
      tags:
        - Admin
      deprecated: true
      description: |-
        **Deprecated in favour of `PUT /_config`**
        This is for replacing all the log keys and optionally changing the console log level.
      summary: Replace console log keys and change log level
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: A map of the console log keys to replace all the current log keys with.
              properties:
                log_key:
                  type: boolean
                  description: The name of the log key and the value `true` to use it.
        description: The map of log keys to use for console logging.
      parameters:
        - $ref: '#/components/parameters/log-level'
        - $ref: '#/components/parameters/log-level-int'
    post:
      responses:
        '200':
          description: Log keys successfully upserted.
        '400':
          $ref: '#/components/responses/request-problem'
      tags:
        - Admin
      deprecated: true
      summary: Update console log keys and change log level
      description: |-
        **Deprecated in favour of `PUT /_config`**
        This is for upserting the log keys provided and optionally changing the console log level.
      parameters:
        - $ref: '#/components/parameters/log-level'
        - $ref: '#/components/parameters/log-level-int'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                log_key:
                  type: boolean
                  description: 'The console log key to upsert along with the value. If the value is `false` then the log key will not be used, or else if `true` than the log key will be used.'
        description: The console log keys to upsert.
  '/_profile/{profilename}':
    parameters:
      - name: profilename
        schema:
          type: string
          enum:
            - heap
            - block
            - threadcreate
            - mutex
            - goroutine
        in: path
        required: true
        description: The handler to use for profiling.
    post:
      responses:
        '200':
          description: Successfully started profiling
        '400':
          $ref: '#/components/responses/request-problem'
        '404':
          $ref: '#/components/responses/Not-found'
      tags:
        - Admin
      description: |-
        This endpoint allows you to start profiling Sync Gateway. Profiling outputs a Go pprof file so performance can be analysed.

        To start profiling, call this endpoint with the handler to use for profiling and supply a file to output the pprof file to.
      summary: Start profiling
  /_profile:
    post:
      responses:
        '200':
          description: Sucessfully started or stopped CPU profiling
        '400':
          $ref: '#/components/responses/request-problem'
      tags:
        - Admin
      description: |-
        This endpoint allows you to start or stop CPU profiling for Sync Gateway. Profiling outputs a Go pprof file so performance can be analysed.

        To start profiling the CPU, call this endpoint and supply a file to output the pprof file to. To stop profiling, call this endpoint but don't supply the `file` in the body.
      summary: Start or stop CPU profiling
      requestBody:
        $ref: '#/components/requestBodies/Profile'
  /_heap:
    post:
      responses:
        '200':
          description: Successfully dumped heap profile
        '400':
          $ref: '#/components/responses/request-problem'
      tags:
        - Admin
      summary: Dump heap profile
      description: This endpoint will dump a pprof heap profile.
      requestBody:
        $ref: '#/components/requestBodies/Profile'
...
  /_status:
    get:
      responses:
        '200':
          description: Returned the status successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/request-problem'
      tags:
        - Admin
      summary: Get the server status
      description: This will retrieve the status of each database and the overall server status.
...
  /_debug/pprof/block:
    get:
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP-Error'
      tags:
        - Admin
      summary: Get the block pprof debug file
      parameters:
        - $ref: '#/components/parameters/debug-profile-seconds'
    post:
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP-Error'
      tags:
        - Admin
      summary: Get the block pprof debug file
      parameters:
        - $ref: '#/components/parameters/debug-profile-seconds'
...
  /_debug/pprof/mutex:
    get:
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP-Error'
      tags:
        - Admin
      summary: Get the mutex pprof debug file
      parameters:
        - $ref: '#/components/parameters/debug-profile-seconds'
    post:
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP-Error'
      tags:
        - Admin
      summary: Get the mutex pprof debug file
      parameters:
        - $ref: '#/components/parameters/debug-profile-seconds'
...
components:
  parameters:
...
    replicationid:
      name: replicationid
      in: path
      required: true
      schema:
        type: string
      description: What replication to target based on it's replication ID.
    replication-active-only:
      name: activeOnly
      in: query
      required: false
      schema:
        type: boolean
        default: 'false'
      description: Only return replications that are actively running (`state=running`).
    replication-local-only:
      name: localOnly
      in: query
      required: false
      schema:
        type: boolean
        default: 'false'
      description: Only return replications that were started on the current Sync Gateway node.
    replication-include-error:
      name: includeError
      in: query
      required: false
      schema:
        type: boolean
        default: 'true'
      description: Include replications that have stopped due to an error (`state=error`).
    replication-include-config:
      name: includeConfig
      in: query
      required: false
      schema:
        type: boolean
        default: 'false'
      description: Include the replication configuration with each replicator status in the response.
...
  responses:
    Not-found:
      description: Resource could not be found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTP-Error'
          examples:
            Database not found:
              value:
                error: not_found
                reason: no such database "invalid-db"
...
    request-problem:
      description: There was a problem with your request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTP-Error'
...
  schemas:
...
    HTTP-Error:
      title: HTTP-Error
      type: object
      properties:
        error:
          type: string
        reason:
          type: string
          description: The error description.
      required:
        - error
        - reason
...